requirements.txt
python-dotenv
requests
vapi-python
asyncio
aiohttp
pydantic
datetime
json5
typing-extensions
